secured backend-frontend communication with password

+++ till-backend/app.js
@@ -51,125 +51,174 @@ app.use((req, res, next) => {
 //   CREATE   //
 ////////////////
 
-app.post("/Staff", (req, res, next) => {
-  const user = new User({
-    ...req.body,
-  });
-  user
-    .save()
-    .then(() => res.status(201).json({ message: "Utilisateur enregistré !" }))
-    .catch((error) => res.status(400).json({ error }));
+app.post("/Staff/:key", (req, res, next) => {
+  if (req.params.key === process.env.API_KEY) {
+    const user = new User({
+      ...req.body,
+    });
+    user
+      .save()
+      .then(() => res.status(201).json({ message: "Utilisateur enregistré !" }))
+      .catch((error) => res.status(400).json({ error }));
+  } else {
+    res.status(400).json({ error: "Incorrect Key" });
+  }
 });
 
-app.post("/Receipt", (req, res, next) => {
-  const receipt = new Receipt({
-    ...req.body,
-  });
-  receipt
-    .save()
-    .then(() => res.status(201).json({ message: "Utilisateur enregistré !" }))
-    .catch((error) => res.status(400).json({ error }));
+app.post("/Receipt/:key", (req, res, next) => {
+  if (req.params.key === process.env.API_KEY) {
+    const receipt = new Receipt({
+      ...req.body,
+    });
+    receipt
+      .save()
+      .then(() => res.status(201).json({ message: "Utilisateur enregistré !" }))
+      .catch((error) => res.status(400).json({ error }));
+  } else {
+    res.status(400).json({ error: "Incorrect Key" });
+  }
 });
 
-app.post("/Item", (req, res, next) => {
-  const item = new Item({
-    ...req.body,
-  });
-  item
-    .save()
-    .then(() => res.status(201).json({ message: "Objet enregistré !" }))
-    .catch((error) => res.status(400).json({ error }));
+app.post("/Item/:key", (req, res, next) => {
+  if (req.params.key === process.env.API_KEY) {
+    const item = new Item({
+      ...req.body,
+    });
+    item
+      .save()
+      .then(() => res.status(201).json({ message: "Objet enregistré !" }))
+      .catch((error) => res.status(400).json({ error }));
+  } else {
+    res.status(400).json({ error: "Incorrect Key" });
+  }
 });
 
-app.post("/Email", (req, res, next) => {
-  var mailOptions = {
-    from: req.body.address,
-    to: req.body.address,
-    subject: req.body.subject,
-    text: req.body.textcontent,
-    html: req.body.htmlcontent,
-  };
+app.post("/Email/:key", (req, res, next) => {
+  if (req.params.key === process.env.API_KEY) {
+    var mailOptions = {
+      from: req.body.address,
+      to: req.body.address,
+      subject: req.body.subject,
+      text: req.body.textcontent,
+      html: req.body.htmlcontent,
+    };
 
-  transporter.sendMail(mailOptions, function (error, info) {
-    if (error) {
-      res.status(400).json({ error });
-    } else {
-      res.status(201).json({ message: "E-mail envoyé !" });
-    }
-  });
+    transporter.sendMail(mailOptions, function (error, info) {
+      if (error) {
+        res.status(400).json({ error });
+      } else {
+        res.status(201).json({ message: "E-mail envoyé !" });
+      }
+    });
+  } else {
+    res.status(400).json({ error: "Incorrect Key" });
+  }
 });
 
 //////////////
 //   READ   //
 //////////////
 
-app.get("/Staff", (req, res, next) => {
-  User.find()
-    .then((users) => res.status(200).json(users))
-    .catch((error) => res.status(400).json({ error }));
+app.get("/Staff/:key", (req, res, next) => {
+  if (req.params.key === process.env.API_KEY) {
+    User.find()
+      .then((users) => res.status(200).json(users))
+      .catch((error) => res.status(400).json({ error }));
+  } else {
+    res.status(400).json({ error: "Incorrect Key" });
+  }
 });
 
-app.get("/Staff-:name", (req, res, next) => {
-  User.findOne({ name: req.params.name })
-    .then((user) => res.status(200).json(user))
-    .catch((error) => res.status(404).json({ error }));
+app.get("/Items/:key", (req, res, next) => {
+  if (req.params.key === process.env.API_KEY) {
+    Item.find()
+      .then((items) => res.status(200).json(items))
+      .catch((error) => res.status(400).json({ error }));
+  } else {
+    res.status(400).json({ error: "Incorrect Key" });
+  }
 });
 
-app.get("/Items", (req, res, next) => {
-  Item.find()
-    .then((items) => res.status(200).json(items))
-    .catch((error) => res.status(400).json({ error }));
+app.get("/Contact/:key", (req, res, next) => {
+  if (req.params.key === process.env.API_KEY) {
+    Contact.find()
+      .then((items) => res.status(200).json(items))
+      .catch((error) => res.status(400).json({ error }));
+  } else {
+    res.status(400).json({ error: "Incorrect Key" });
+  }
 });
 
-app.get("/Contact", (req, res, next) => {
-  Contact.find()
-    .then((items) => res.status(200).json(items))
-    .catch((error) => res.status(400).json({ error }));
-});
-
-app.get("/Receipts", (req, res, next) => {
-  Receipt.find()
-    .then((receipts) => res.status(200).json(receipts))
-    .catch((error) => res.status(400).json({ error }));
+app.get("/Receipts/:key", (req, res, next) => {
+  if (req.params.key === process.env.API_KEY) {
+    Receipt.find()
+      .then((receipts) => res.status(200).json(receipts))
+      .catch((error) => res.status(400).json({ error }));
+  } else {
+    res.status(400).json({ error: "Incorrect Key" });
+  }
 });
 
 ////////////////
 //   UPDATE   //
 ////////////////
 
-app.put("/Staff/:id", (req, res, next) => {
-  User.updateOne({ _id: req.params.id }, { ...req.body, _id: req.params.id })
-    .then(() => res.status(200).json({ message: "Utilisateur modifié !" }))
-    .catch((error) => res.status(400).json({ error }));
+app.put("/Staff/:id/:key", (req, res, next) => {
+  if (req.params.key === process.env.API_KEY) {
+    User.updateOne({ _id: req.params.id }, { ...req.body, _id: req.params.id })
+      .then(() => res.status(200).json({ message: "Utilisateur modifié !" }))
+      .catch((error) => res.status(400).json({ error }));
+  } else {
+    res.status(400).json({ error: "Incorrect Key" });
+  }
 });
 
-app.put("/Contact/:id", (req, res, next) => {
-  Contact.updateOne({ _id: req.params.id }, { ...req.body, _id: req.params.id })
-    .then(() =>
-      res.status(200).json({ message: "Données de contact modifié !" })
+app.put("/Contact/:id/:key", (req, res, next) => {
+  if (req.params.key === process.env.API_KEY) {
+    Contact.updateOne(
+      { _id: req.params.id },
+      { ...req.body, _id: req.params.id }
     )
-    .catch((error) => res.status(400).json({ error }));
+      .then(() =>
+        res.status(200).json({ message: "Données de contact modifié !" })
+      )
+      .catch((error) => res.status(400).json({ error }));
+  } else {
+    res.status(400).json({ error: "Incorrect Key" });
+  }
 });
 
-app.put("/Item/:id", (req, res, next) => {
-  Item.updateOne({ _id: req.params.id }, { ...req.body, _id: req.params.id })
-    .then(() => res.status(200).json({ message: "Objet modifié !" }))
-    .catch((error) => res.status(400).json({ error }));
+app.put("/Item/:id/:key", (req, res, next) => {
+  if (req.params.key === process.env.API_KEY) {
+    Item.updateOne({ _id: req.params.id }, { ...req.body, _id: req.params.id })
+      .then(() => res.status(200).json({ message: "Objet modifié !" }))
+      .catch((error) => res.status(400).json({ error }));
+  } else {
+    res.status(400).json({ error: "Incorrect Key" });
+  }
 });
 
-app.put("/Category/Update", (req, res, next) => {
-  Item.updateMany(
-    { category: req.body.oldCategory },
-    { category: req.body.newCategory }
-  )
-    .then(() => res.status(200).json({ message: "Catégorie modifiée !" }))
-    .catch((error) => res.status(400).json({ error }));
+app.put("/Category/Update/:key", (req, res, next) => {
+  if (req.params.key === process.env.API_KEY) {
+    Item.updateMany(
+      { category: req.body.oldCategory },
+      { category: req.body.newCategory }
+    )
+      .then(() => res.status(200).json({ message: "Catégorie modifiée !" }))
+      .catch((error) => res.status(400).json({ error }));
+  } else {
+    res.status(400).json({ error: "Incorrect Key" });
+  }
 });
 
-app.put("/Receipt/Payement/:id", (req, res, next) => {
-  Receipt.updateOne({ _id: req.params.id }, { payement: req.body.payement })
-    .then(() => res.status(200).json({ message: "Ticket modifié !" }))
-    .catch((error) => res.status(400).json({ error }));
+app.put("/Receipt/Payement/:id/:key", (req, res, next) => {
+  if (req.params.key === process.env.API_KEY) {
+    Receipt.updateOne({ _id: req.params.id }, { payement: req.body.payement })
+      .then(() => res.status(200).json({ message: "Ticket modifié !" }))
+      .catch((error) => res.status(400).json({ error }));
+  } else {
+    res.status(400).json({ error: "Incorrect Key" });
+  }
 });
 
 ////////////////
@@ -177,15 +226,23 @@ app.put("/Receipt/Payement/:id", (req, res, next) => {
 ////////////////
 
 app.delete("/Staff-:id", (req, res, next) => {
-  User.deleteOne({ _id: req.params.id })
-    .then(() => res.status(200).json({ message: "User removed !" }))
-    .catch((error) => res.status(400).json({ error }));
+  if (req.params.key === process.env.API_KEY) {
+    User.deleteOne({ _id: req.params.id })
+      .then(() => res.status(200).json({ message: "User removed !" }))
+      .catch((error) => res.status(400).json({ error }));
+  } else {
+    res.status(400).json({ error: "Incorrect Key" });
+  }
 });
 
 app.delete("/Item-:id", (req, res, next) => {
-  Item.deleteOne({ _id: req.params.id })
-    .then(() => res.status(200).json({ message: "User removed !" }))
-    .catch((error) => res.status(400).json({ error }));
+  if (req.params.key === process.env.API_KEY) {
+    Item.deleteOne({ _id: req.params.id })
+      .then(() => res.status(200).json({ message: "User removed !" }))
+      .catch((error) => res.status(400).json({ error }));
+  } else {
+    res.status(400).json({ error: "Incorrect Key" });
+  }
 });
 
 module.exports = app;
